+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/FabricPeripheralWorksClient.kt
@@ -22,7 +22,7 @@ object FabricPeripheralWorksClient : ClientModInitializer {
         }
         ModelLoadingRegistry.INSTANCE.registerResourceProvider { ExpansionModelProvider }
 
-        BlockRenderLayerMap.INSTANCE.putBlock(Blocks.FLEXIBLE_REALITY_ANCHOR.get(), RenderType.cutout())
+        BlockRenderLayerMap.INSTANCE.putBlock(Blocks.FLEXIBLE_REALITY_ANCHOR.get(), RenderType.translucent())
         BlockRenderLayerMap.INSTANCE.putBlock(Blocks.FLEXIBLE_STATUE.get(), RenderType.translucent())
     }
 }