+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.2] - 2023-08-01
+
+### Added
+
+- Reality forger `batchForgeRealityPieces`
+
 ### Changed
 
 - Flexible reality anchor now use translucent